+++ CHANGELOG.md
@@ -90,3 +90,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased] - 2022-04-23
 ### Added
 - Added new plotting script plot_num_subclones_box.R
+- Added plot_accessories.R
+
+### Changed
+- Moved seed, colour scheme and legends to plot_accesories.R